fix: DTO 변수명 변경
