+++ AIProject/AIProject/Data/Response/UpbitResponse/CoinDTO.swift
@@ -1,31 +1,25 @@
 //
-//  CoinDTO.swift
+//  Response.swift
 //  AIProject
 //
-//  Created by 강대훈 on 7/31/25.
+//  Created by kangho lee on 7/29/25.
 //
 
 import Foundation
 
-/// WebSocket으로부터 오는 실시간 시세 데이터 DTO
-struct CoinDTO: Decodable {
-    /// 종목 코드
-    let code: String
-    /// 현재 체결 가격
-    let tradePrice: Double
-    /// 전일 대비 변화 (RISE, FALL, EVEN)
-    let change: String
-    /// 전일 대비 가격 변화
-    let changePrice: Double
-    /// 전일 대비 등락률
-    let changeRate: Double
+/// 업비트에서 거래 가능한 종목 목록 DTO
+struct CoinDTO: Codable {
+    /// 마켓 식별자
+    let coinID: String
+    /// 코인의 한글 이름 ex) 비트코인
+    let koreanName: String
+    /// 코인의 영문 이름 ex) Bitcoin
+    let englishName: String
 
     enum CodingKeys: String, CodingKey {
-        case code
-        case tradePrice = "trade_price"
-        case change
-        case changePrice = "change_price"
-        case changeRate = "change_rate"
+        case coinID = "market"
+        case koreanName = "korean_name"
+        case englishName = "english_name"
     }
 }
