+++ AIProject/AIProject/Data/Response/UpbitResponse/TickerDTO.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 /// 특정 코인의 실시간 시세, 다양한 시세 관련 정보를 포함하는 데이터 DTO
-struct CoinQuoteDTO: Codable {
+struct TickerDTO: Codable {
     /// 코인 마켓 코드
     let market: String
     /// 최근 체결 일자 및 시각
@@ -77,7 +77,7 @@ struct CoinQuoteDTO: Codable {
     }
 }
 
-extension CoinQuoteDTO {
+extension TickerDTO {
     /// 체결 일자 및 시간을 Date 형식으로 반환합니다.
     var tradeDateTime: Date {
         Date(timeIntervalSince1970: TimeInterval(tradeTimestamp) / 1000)
